requirements.txt
Flask==2.0.2
python-dotenv==0.19.1
Flask-WTF==0.15.1